change few output string and make timeout as double rather than long

+++ NmapOps.h
@@ -390,7 +390,7 @@ class NmapOps {
   int scripttrace;
   int scriptupdatedb;
   bool scripthelp;
-  long scripttimeout;
+  double scripttimeout;
   void chooseScripts(char* argument);
   std::vector<std::string> chosenScripts;
 #endif
